refactor: /community/[boardCode] 페이지 전환
